requirements.txt
streamlit
ultralytics
opencv-python-headless
numpy
filterpy
scikit-image
pillow